Reworked constructors for StoryGraph, PageNode, ChildLink objects
Began testing StoryGraph, PageNode, and ChildLink objects and discovered issues with how the constructors would take input data. Nick and I did a lot of work to streamline with the various classes expect as input data. We can now convert from json data to Python objects and back again, which allows us to use an object oriented approach to tree edits and updates.