More testing of Editor.js. Added duplicateStory() function.
Deleting nodes from stories tested with basic operations and a handful of edge cases including deleting the root. Fixed an issue where parents of deleted node still listed the deleted node as a child. Added duplicateStory() function to Editor class which uses a timestamp to differentiate the new copy. Added a MAX_SIZE field to StoryStack so the undo history is limited to 99.